Extract date/time parts
